test files now separated
